// src/types/types.ts
export interface ICountry {
  flagUrlImg: string;
  commonName: string;
  nativeName: string;
  population: number;
  region: string;
  subregion?: string;
  capital?: string[];
  tld?: string[];
  currencies?: string[];
  languages?: string[];
  borders?: string[];
  cca3: string;
}

export interface ICountryApiResponse {
  name: {
    common: string;
    official: string;
    nativeName?: {
      [key: string]: {
        official: string;
        common: string;
      };
    };
  };
  tld?: string[];
  cca2: string;
  ccn3?: string;
  cca3: string;
  independent?: boolean;
  status: string;
  unMember: boolean;
  currencies?: {
    [key: string]: {
      name: string;
      symbol: string;
    };
  };
  idd: {
    root?: string;
    suffixes?: string[];
  };
  capital?: string[];
  altSpellings: string[];
  region: string;
  subregion?: string;
  languages?: {
    [key: string]: string;
  };
  translations: {
    [key: string]: {
      official: string;
      common: string;
    };
  };
  latlng: number[];
  landlocked: boolean;
  area: number;
  demonyms?: {
    [key: string]: {
      f: string;
      m: string;
    };
  };
  flag: string;
  maps: {
    googleMaps: string;
    openStreetMaps: string;
  };
  population: number;
  car: {
    signs?: string[];
    side: string;
  };
  timezones: string[];
  continents: string[];
  flags: {
    png: string;
    svg: string;
    alt?: string;
  };
  coatOfArms: {
    png?: string;
    svg?: string;
  };
  startOfWeek: string;
  capitalInfo: {
    latlng?: number[];
  };
  postalCode?: {
    format: string;
    regex?: string;
  };
  borders?: string[];
  cioc?: string;
  gini?: {
    [key: string]: number;
  };
  fifa?: string;
}

export interface CountriesState {
  status: 'loading' | 'succeeded' | 'failed' | 'idle';
  data: ICountry[];
  error: string | null;
}

export interface IRegionOption {
  value: string;
  label: string;
}